+++ apps/frontend/src/index.tsx
@@ -14,6 +14,7 @@ import { store } from "~/libs/modules/store/store.js";
 import { AccessManagement } from "~/pages/access-management/access-management.jsx";
 import { Analytics } from "~/pages/analytics/analytics.jsx";
 import { Auth } from "~/pages/auth/auth.jsx";
+import { Contributors } from "~/pages/contributors/contributors.jsx";
 import { NoAccess } from "~/pages/no-access/no-access.jsx";
 import { NotFound } from "~/pages/not-found/not-found.jsx";
 import { Profile } from "~/pages/profile/profile.jsx";
@@ -56,6 +57,14 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 									</ProtectedRoute>
 								),
 								path: AppRoute.PROFILE,
+							},
+							{
+								element: (
+									<ProtectedRoute>
+										<Contributors />
+									</ProtectedRoute>
+								),
+								path: AppRoute.CONTRIBUTORS,
 							},
 							{
 								element: (